organization/requirements.txt
Django==3.1.4
psycopg2==2.8.6
djangorestframework==3.12.2
drf-jwt==1.17.3
django-cors-middleware==1.3.1